+++ tests/checkpointing/test_model_checkpoint.py
@@ -13,33 +13,31 @@
 # limitations under the License.
 import os
 import os.path as osp
-import pytorch_lightning as pl
-from distutils.version import LooseVersion
-from unittest.mock import MagicMock, Mock
-
-import yaml
 import pickle
 import platform
 import re
 from pathlib import Path
+from unittest import mock
+from unittest.mock import Mock
 
 import cloudpickle
 import pytest
 import torch
+import yaml
 
+import pytorch_lightning as pl
 import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer, seed_everything
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import TensorBoardLogger
-from tests.base import EvalModelTemplate, BoringModel
 from pytorch_lightning.utilities.cloud_io import load as pl_load
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from tests.base import BoringModel, EvalModelTemplate
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 @pytest.mark.parametrize('save_top_k', [-1])
 def test_model_checkpoint_correct_score(tmpdir, save_top_k):
-    os.environ['PL_DEV_DEBUG'] = '1'
-
     """Test that when a model checkpoint is saved, it saves with the correct score appended to ckpt_path"""
     tutils.reset_seed()
 
@@ -413,10 +411,10 @@ def test_ckpt_metric_names(tmpdir):
     assert len(val) > 3
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_default_checkpoint_behavior(tmpdir):
     seed_everything(1234)
 
-    os.environ['PL_DEV_DEBUG'] = '1'
     model = EvalModelTemplate()
     model.validation_step = model.validation_step_no_monitor
     model.validation_epoch_end = model.validation_epoch_end_no_monitor
@@ -529,9 +527,9 @@ def test_model_checkpoint_save_last_checkpoint_contents(tmpdir):
         assert w0.eq(w1).all()
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 @pytest.mark.parametrize('mode', ['min', 'max'])
 def test_checkpointing_with_nan_as_first(tmpdir, mode):
-    os.environ['PL_DEV_DEBUG'] = '1'
     monitor = [float('nan')]
     monitor += [5, 7, 8] if mode == 'max' else [8, 7, 5]
 
@@ -554,12 +552,11 @@ def test_checkpointing_with_nan_as_first(tmpdir, mode):
     assert trainer.dev_debugger.checkpoint_callback_history[-1]['epoch'] == len(monitor) - 1
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_checkpoint_repeated_strategy(tmpdir):
     """
     This test validates that the checkpoint can be called when provided to callacks list
     """
-
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     checkpoint_callback = ModelCheckpoint(monitor='val_loss', dirpath=tmpdir, filename="{epoch:02d}")
 
@@ -605,12 +602,11 @@ def test_checkpoint_repeated_strategy(tmpdir):
         assert str(os.listdir(tmpdir)) == "['epoch=00.ckpt']"
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_checkpoint_repeated_strategy_tmpdir(tmpdir):
     """
     This test validates that the checkpoint can be called when provided to callacks list
     """
-
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     checkpoint_callback = ModelCheckpoint(monitor='val_loss', filepath=os.path.join(tmpdir, "{epoch:02d}"))
 
@@ -661,13 +657,12 @@ def test_checkpoint_repeated_strategy_tmpdir(tmpdir):
         assert sorted(os.listdir(path_to_lightning_logs)) == sorted([f'version_{i}' for i in range(idx + 1)])
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_checkpoint_repeated_strategy_extended(tmpdir):
     """
     This test validates checkpoint can be called several times without
     increasing internally its global step if nothing run.
     """
-
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class ExtendedBoringModel(BoringModel):