+++ tests/trainer/data_flow/test_train_loop_flow_scalar_1_0.py
@@ -14,20 +14,23 @@
 """
 Tests to ensure that the training loop works with a dict (1.0)
 """
-from pytorch_lightning.core.lightning import LightningModule
-import pytest
-from pytorch_lightning import Trainer
-from tests.base.deterministic_model import DeterministicModel
-from tests.base.boring_model import BoringModel
 import os
+from unittest import mock
+
+import pytest
 import torch
 
+from pytorch_lightning import Trainer
+from pytorch_lightning.core.lightning import LightningModule
+from tests.base.boring_model import BoringModel
+from tests.base.deterministic_model import DeterministicModel
 
+
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__flow_scalar(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -58,11 +61,11 @@ def test__training_step__flow_scalar(tmpdir):
     assert not model.training_epoch_end_called
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__tr_step_end__flow_scalar(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -100,11 +103,11 @@ def test__training_step__tr_step_end__flow_scalar(tmpdir):
     assert not model.training_epoch_end_called
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__epoch_end__flow_scalar(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -148,11 +151,11 @@ def test__training_step__epoch_end__flow_scalar(tmpdir):
     assert model.training_epoch_end_called
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__step_end__epoch_end__flow_scalar(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):