+++ tests/trainer/logging_tests/test_eval_loop_logging_1_0.py
@@ -14,21 +14,23 @@
 """
 Tests to ensure that the training loop works with a dict (1.0)
 """
-from pytorch_lightning.core.lightning import LightningModule
-from pytorch_lightning import Trainer
-from pytorch_lightning import callbacks, seed_everything
-from tests.base.deterministic_model import DeterministicModel
-from tests.base import SimpleModule, BoringModel
 import os
-import torch
+from unittest import mock
+
 import pytest
+import torch
+
+from pytorch_lightning import Trainer, callbacks, seed_everything
+from pytorch_lightning.core.lightning import LightningModule
+from tests.base import BoringModel, SimpleModule
+from tests.base.deterministic_model import DeterministicModel
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__validation_step__log(tmpdir):
     """
     Tests that validation_step can log
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -84,11 +86,11 @@ def test__validation_step__log(tmpdir):
     assert expected_cb_metrics == callback_metrics
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__validation_step__step_end__epoch_end__log(tmpdir):
     """
     Tests that validation_step can log
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -162,12 +164,12 @@ def test__validation_step__step_end__epoch_end__log(tmpdir):
     assert expected_cb_metrics == callback_metrics
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 @pytest.mark.parametrize(['batches', 'log_interval', 'max_epochs'], [(1, 1, 1), (64, 32, 2)])
 def test_eval_epoch_logging(tmpdir, batches, log_interval, max_epochs):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(BoringModel):
         def validation_epoch_end(self, outputs):
@@ -211,11 +213,11 @@ def test_eval_epoch_logging(tmpdir, batches, log_interval, max_epochs):
     assert len(trainer.dev_debugger.logged_metrics) == max_epochs
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_eval_float_logging(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(BoringModel):
 
@@ -246,13 +248,12 @@ def test_eval_float_logging(tmpdir):
     assert logged_metrics == expected_logged_metrics
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_eval_logging_auto_reduce(tmpdir):
     """
     Tests that only training_step can be used
     """
     seed_everything(1234)
-
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(BoringModel):
         def on_pretrain_routine_end(self) -> None:
@@ -315,12 +316,12 @@ def test_eval_logging_auto_reduce(tmpdir):
     assert len(logged_val) == 6
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 @pytest.mark.parametrize(['batches', 'log_interval', 'max_epochs'], [(1, 1, 1), (64, 32, 2)])
 def test_eval_epoch_only_logging(tmpdir, batches, log_interval, max_epochs):
     """
     Tests that only test_epoch_end can be used to log, and we return them in the results.
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(BoringModel):
         def test_epoch_end(self, outputs):