+++ tests/trainer/logging_tests/test_train_loop_logging_1_0.py
@@ -15,6 +15,7 @@
 Tests to ensure that the training loop works with a dict (1.0)
 """
 import os
+from unittest import mock
 
 import pytest
 import torch
@@ -26,11 +27,11 @@ from tests.base.boring_model import BoringModel, RandomDictDataset, RandomDictSt
 from tests.base.deterministic_model import DeterministicModel
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__log(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -122,11 +123,11 @@ def test__training_step__log(tmpdir):
     assert callback_metrics == expected_callback_metrics
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test__training_step__epoch_end__log(tmpdir):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(DeterministicModel):
         def training_step(self, batch, batch_idx):
@@ -184,12 +185,12 @@ def test__training_step__epoch_end__log(tmpdir):
     assert callback_metrics == expected_callback_metrics
 
 
+@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 @pytest.mark.parametrize(['batches', 'log_interval', 'max_epochs'], [(1, 1, 1), (64, 32, 2)])
 def test__training_step__step_end__epoch_end__log(tmpdir, batches, log_interval, max_epochs):
     """
     Tests that only training_step can be used
     """
-    os.environ['PL_DEV_DEBUG'] = '1'
 
     class TestModel(BoringModel):
         def training_step(self, batch, batch_idx):